When writing a box to disk, do not delete the 'packed' metadata.
This will have the effect of fixing the file mtime at the moment at which we
start to use tools to manipulate the metadata, on the assumption that at this
point the box is pretty much packed, and we don't need to update the packed
timestamp, even if we continue to manipulate the metadata. If we want to update
the packed timestamp, we could delete or edit the entry in the header.